Changed output CSV, added Groups of Four (Beta) option
The output CSV now has separate columns for the group number and satisfaction percentage so that they can be sorted better.

Also, a new option for Groups of Four has been incorporated. This will lead to better teams overall, but will take a large amount of time (from 1 to 2 hours, perhaps) and a significant percentage of system resources. (I've only run it on a system with 8GB RAM, can't be sure it'll run on one with 4GB!)